+++ foo_ui_columns/ng_playlist/ng_playlist_artwork.h
@@ -11,16 +11,29 @@ enum class ArtworkReaderStatus {
     Aborted,
 };
 
+struct ArtworkRenderingContext {
+    using Ptr = std::shared_ptr<ArtworkRenderingContext>;
+
+    wil::com_ptr<ID3D11Device> d3d_device;
+    wil::com_ptr<ID2D1Factory1> d2d_factory;
+    wil::com_ptr<ID2D1Device> d2d_device;
+    wil::com_ptr<ID2D1DeviceContext> d2d_device_context;
+    wil::com_ptr<IWICImagingFactory> wic_factory;
+
+    static Ptr s_create(unsigned width, unsigned height);
+
+    void enlarge(unsigned width, unsigned height) const;
+};
+
 class ArtworkReader {
 public:
     using Ptr = std::shared_ptr<ArtworkReader>;
 
-    ArtworkReader(metadb_handle_ptr track, int width, int height, COLORREF background_colour, bool show_reflection,
+    ArtworkReader(metadb_handle_ptr track, int width, int height, bool show_reflection,
         OnArtworkLoadedCallback callback, std::shared_ptr<class ArtworkReaderManager> manager)
         : m_handle(std::move(track))
         , m_width(width)
         , m_height(height)
-        , m_background_colour(background_colour)
         , m_show_reflection(show_reflection)
         , m_callback(std::move(callback))
         , m_manager(std::move(manager))
@@ -72,7 +85,7 @@ public:
         m_callback(this);
     }
 
-    void start();
+    void start(ArtworkRenderingContext::Ptr context);
 
     void wait()
     {
@@ -81,14 +94,16 @@ public:
     }
 
 private:
-    unsigned read_artwork(abort_callback& p_abort);
+    static constexpr GUID artwork_type_id = album_art_ids::cover_front;
+
+    album_art_data_ptr read_artwork(abort_callback& p_abort);
+    void render_artwork(const ArtworkRenderingContext::Ptr& context, album_art_data_ptr data, abort_callback& p_abort);
 
     std::unordered_map<GUID, wil::shared_hbitmap> m_bitmaps;
     metadb_handle_ptr m_handle;
-    int m_width{0};
-    int m_height{0};
-    COLORREF m_background_colour{RGB(255, 255, 255)};
-    bool m_show_reflection{false};
+    int m_width{};
+    int m_height{};
+    bool m_show_reflection{};
     OnArtworkLoadedCallback m_callback;
     std::shared_ptr<class ArtworkReaderManager> m_manager;
     abort_callback_impl m_abort;
@@ -115,8 +130,8 @@ public:
 
     void reset() { abort_current_tasks(); }
 
-    void request(const metadb_handle_ptr& p_handle, ArtworkReader::Ptr& p_out, int cx, int cy, COLORREF cr_back,
-        bool b_reflection, OnArtworkLoadedCallback p_notify);
+    void request(const metadb_handle_ptr& p_handle, ArtworkReader::Ptr& p_out, int cx, int cy, bool b_reflection,
+        OnArtworkLoadedCallback p_notify);
 
     void flush_pending()
     {
@@ -125,7 +140,9 @@ public:
 
         auto reader = *m_pending_readers.rbegin();
         m_pending_readers.pop_back();
-        reader->start();
+
+        auto context = get_d2d_device_context();
+        reader->start(std::move(context));
         m_current_readers.emplace_back(std::move(reader));
     }
 
@@ -137,13 +154,13 @@ public:
         m_nocover_bitmap.reset();
     }
 
-    void on_reader_done(const ArtworkReader* ptr);
+    void on_reader_done(ArtworkRenderingContext::Ptr context, const ArtworkReader* ptr);
 
     ArtworkReaderManager() = default;
 
     // Called from reader worker thread
     wil::shared_hbitmap request_nocover_image(
-        int cx, int cy, COLORREF cr_back, bool b_reflection, abort_callback& p_abort);
+        const ArtworkRenderingContext::Ptr& context, int cx, int cy, bool b_reflection, abort_callback& p_abort);
 
     void flush_nocover()
     {
@@ -154,6 +171,11 @@ public:
 private:
     static constexpr size_t max_readers{4};
 
+    std::shared_ptr<ArtworkRenderingContext> get_d2d_device_context();
+
+    wil::com_ptr<ID2D1Factory1> m_d2d_factory;
+    std::vector<std::shared_ptr<ArtworkRenderingContext>> m_rendering_contexts;
+
     std::vector<ArtworkReader::Ptr> m_aborting_readers;
     std::vector<ArtworkReader::Ptr> m_current_readers;
     std::vector<ArtworkReader::Ptr> m_pending_readers;
@@ -163,5 +185,7 @@ private:
     size_t m_nocover_cx{};
     size_t m_nocover_cy{};
 };
+
+bool get_default_artwork_placeholder_data(album_art_data_ptr& p_out, abort_callback& p_abort);
 
 } // namespace cui::panels::playlist_view